fix watch timeout check
When a dynamic client watcher is used, timeout_seconds is always present
in self.func.keywords, it is set in DynamicClient.watch(). Its value is
None when the argument was not passed by the user.

The current check done on TimeoutError when awaiting from
self.resp.content.readline() prevents to reopen the request in this
case.

Fix the check for timeout_seconds presence. If it is None or 0, it will
now behave as if the 'timeout_seconds' key is not present.

Signed-off-by: Olivier Matz <olivier.matz@6wind.com>