save command return code in ws_client
When the websocket API is used to execute a command on a pod, the status is sent
over the ERROR_CHANNEL on termination. Parse this information and set a
"returncode" field to the exit code of the command.

Update the pod_exec example to make use of this new field.

Signed-off-by: Olivier Matz <olivier.matz@6wind.com>